+++ package.json
@@ -39,7 +39,7 @@
     "shelljs": "^0.7.0",
     "source-map-support": "^0.4.17",
     "time-grunt": "1.3.0",
-    "tslib": "^1.7.1",
+    "tslib": "^1.9.3",
     "tslint": "^5.4.3",
     "typedoc": "^0.5.10",
     "typedoc-plugin-external-module-name": "git://github.com/PanayotCankov/typedoc-plugin-external-module-name.git#with-js",
