+++ tns-core-modules/globals/globals.ts
@@ -1,18 +1,5 @@
 ﻿// Required by TypeScript compiler
-require("./decorators");
-
-// Required by V8 snapshot generator
-if (!global.__extends) {
-    global.__extends = function (d, b) {
-        for (var p in b) {
-            if (b.hasOwnProperty(p)) {
-                d[p] = b[p];
-            }
-        }
-        function __() { this.constructor = d; }
-        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-    };
-}
+require("./ts-helpers");
 
 // This method iterates all the keys in the source exports object and copies them to the destination exports one.
 // Note: the method will not check for naming collisions and will override any already existing entries in the destination exports.
