+++ tns-core-modules/package.json
@@ -26,7 +26,8 @@
   "license": "Apache-2.0",
   "typings": "tns-core-modules.d.ts",
   "dependencies": {
-    "tns-core-modules-widgets": "next"
+    "tns-core-modules-widgets": "next",
+    "tslib": "^1.9.3"
   },
   "devDependencies": {
     "@types/node": "~7.0.5",
